requirements.txt
Django==2.0.6
djangorestframework==3.8.2
mysqlclient==1.3.14
django-mysql==2.3.0
requests==2.18.4
django-filter==1.1.0
Faker==0.9.0
